+++ src/main/java/com/edsonmoreirajr/votacao/gateway/impl/PautaGatewayImpl.java
@@ -1,9 +1,11 @@
 package com.edsonmoreirajr.votacao.gateway.impl;
 
+import com.edsonmoreirajr.votacao.config.message.MessageSourceService;
 import com.edsonmoreirajr.votacao.dto.TotalVotosDto;
 import com.edsonmoreirajr.votacao.entity.Pauta;
 import com.edsonmoreirajr.votacao.gateway.PautaGateway;
 import com.edsonmoreirajr.votacao.repository.PautaRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -16,7 +18,8 @@ import java.util.Optional;
 @Component
 @RequiredArgsConstructor
 public class PautaGatewayImpl implements PautaGateway {
-
+    private final static String ENTITY_NOT_FOUND_PAUTA = "entity-not-found.pauta";
+    private final MessageSourceService messageSourceService;
     private final PautaRepository pautaRepository;
 
     @Override
@@ -36,6 +39,10 @@ public class PautaGatewayImpl implements PautaGateway {
 
     @Override
     public void deletePauta(Long id) {
+        var pauta = pautaRepository.findById(id);
+        if (pauta.isEmpty()) {
+            throw new EntityNotFoundException(messageSourceService.getMessage(ENTITY_NOT_FOUND_PAUTA, id));
+        }
         pautaRepository.deleteById(id);
     }
